requirements.txt
google-genai
websockets
python-dotenv
pyttsx3
asyncio
requests
gtts